ejemplo hashmap commmit
