+++ src/main/java/org/example/lisarraylist/Empleado.java
@@ -43,27 +43,6 @@ public class Empleado extends Persona {
     }
 
 
-    public void ejecutar() {
-
-
-        ArrayList<Persona> personas = new ArrayList<>();
-        personas.add(new Persona("Juan", "Perez", "1233445", "juand@gmail.com"));
-        personas.add(new Persona("Pedro", "Perez", "1233445", "juand@gmail.com"));
-        personas.add(new Empleado("yeison", "Osorio", "1233445", "yeisond@gmail.com", 50000));
-        personas.add(new Empleado("Fernado", "Rodriguez", "1233445", "Fernado@gmail.com", 80000));
-
-
-        System.out.println("Personas y empleados registrados:");
-        for (Persona persona : personas) {
-            System.out.print(persona.getNombre() + " " + persona.getApellido() + " " + persona.getDocumento() + " " + persona.getCorreo());
-            ;
-            if (persona instanceof Empleado) {
-                System.out.println(":" + "el sueldo es: " + " ($" + ((Empleado) persona).getSueldo() + ")");
-            } else {
-                System.out.println();
-            }
-        }
-    }
 
 
 }
