+++ src/main/java/org/example/map/Animal.java
@@ -1,0 +1,30 @@
+package org.example.map;
+
+public class Animal {
+
+
+    private String nombre;
+
+    private String especie;
+
+    public Animal(String nombre, String especie) {
+        this.nombre = nombre;
+        this.especie = especie;
+    }
+
+    public String getNombre() {
+        return nombre;
+    }
+
+    public void setNombre(String nombre) {
+        this.nombre = nombre;
+    }
+
+    public String getEspecie() {
+        return especie;
+    }
+
+    public void setEspecie(String especie) {
+        this.especie = especie;
+    }
+}
