+++ src/main/java/org/example/map/Perro.java
@@ -1,0 +1,12 @@
+package org.example.map;
+
+class Perro extends Animal {
+
+    private String raza;
+
+
+    public Perro(String nombre, String especie, String raza) {
+        super(nombre, especie);
+        this.raza = raza;
+    }
+}
