+++ src/main/java/org/example/map/Pez.java
@@ -1,0 +1,11 @@
+package org.example.map;
+
+public class Pez extends Animal {
+
+    private String raza;
+
+    public Pez(String nombre, String especie, String raza) {
+        super(nombre, especie);
+        this.raza = raza;
+    }
+}
